+++ plookup/src/table/lookup_table.rs
@@ -179,6 +179,12 @@ impl PlookupTable3Arity {
 /// This is a table, either
 pub struct PlookupTable4Arity(Vec<[BlsScalar; 4]>);
 
+impl Default for PlookupTable4Arity {
+    fn default() -> Self {
+        PlookupTable4Arity::new()
+    }
+}
+
 impl PlookupTable4Arity {
     /// Create a new, empty Plookup table, with arity 4.
     pub fn new() -> Self {
@@ -313,8 +319,8 @@ impl PlookupTable4Arity {
     }
 
     /// Attempts to find an output value, given two input values, by querying the lookup
-    /// table. The final wire holds the index of the table. The element must be predetermined 
-    /// to be between -1 and 2 depending on the type of table used. 
+    /// table. The final wire holds the index of the table. The element must be predetermined
+    /// to be between -1 and 2 depending on the type of table used.
     /// If the element does not exist, it will return an error.
     pub fn lookup(
         &self,