+++ CHANGELOG.md
@@ -7,13 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
-- Changed `compute_permutation_poly` to simpler version (#350)
 - Refactored to/from_bytes criteria for some structs (#333)
+
+## [0.3.6] - 17-12-20
+### Added 
+- To/From bytes impl for `PublicInput`.
+### Changed
+- Changed `compute_permutation_poly` to simpler version.
 
 ## [0.3.5] - 25-11-20
 ### Changed
 - Changed `Proof` & `ProofEvaluations` byte conversion fn signatures.
-
 ### Added
 - Implemented `Canon` for `Proof`.